test_service/requirements.txt
fastapi
uvicorn
pymongo
requests
python-multipart
prometheus_fastapi_instrumentator
opentelemetry-distro[otlp]
opentelemetry-instrumentation